+++ .cargo/config.toml
@@ -4,7 +4,8 @@ target = "wasm32-unknown-unknown"
 [target.wasm32-unknown-unknown]
 rustflags = [
     # Import memory from WASM-4
-    "-C", "link-arg=--import-memory",
+    "-C",
+    "link-arg=--import-memory",
     # "-C", "link-arg=--initial-memory=65536",
     # "-C", "link-arg=--max-memory=65536",
 
@@ -12,7 +13,8 @@ rustflags = [
     # Reserve 8192 bytes of Rust stack space, offset from 6560.
     # Bump this value, 16-byte aligned, if the framebuffer gets corrupted.
     # "-C", "link-arg=-zstack-size=14752",
-    "-C", "link-arg=-zstack-size=10656",
+    "-C",
+    "link-arg=-zstack-size=10656",
     # "-C", "link-arg=-zstack-size=22944",
 ]
