+++ src/alloc.rs
@@ -3,7 +3,7 @@ use buddy_alloc::{BuddyAllocParam, FastAllocParam, NonThreadsafeAlloc};
 // These values can be tuned
 const FAST_HEAP_SIZE: usize = 4 * 1024; // 4 KB
 const HEAP_SIZE: usize = 24 * 1024; // 24 KB
-const LEAF_SIZE: usize = 16;
+const LEAF_SIZE: usize = 32; // 16 B
 
 static mut FAST_HEAP: [u8; FAST_HEAP_SIZE] = [0u8; FAST_HEAP_SIZE];
 static mut HEAP: [u8; HEAP_SIZE] = [0u8; HEAP_SIZE];
