+++ src/common_types.rs
@@ -5,16 +5,6 @@ pub struct Coord {
     pub x: f64,
     pub y: f64,
 }
-
-// Trait for anything that can move on the screen
-pub trait Movable {
-    fn update_position(&mut self);
-}
-// Trait for anything that's visible on screen
-pub trait Visible {
-    fn draw(&self);
-}
-
 impl Coord {
     /// Absolute distance between two points
     pub fn distance_to(self, other: &Self) -> f64 {
