+++ src/entities/mod.rs
@@ -1,0 +1,6 @@
+pub mod bomb;
+pub mod enemy;
+pub mod entity;
+pub mod manager;
+pub mod player;
+pub(crate) mod traits;
