+++ src/sound/mod.rs
@@ -1,0 +1,3 @@
+pub mod effects;
+pub mod music;
+pub mod notes;
